Move spiral selection in equatorial viewpoint panel
