+++ src/org/helioviewer/jhv/layers/selector/LayersPanel.java
@@ -48,7 +48,7 @@ public class LayersPanel extends JPanel {
     private static final int REMOVE_COL = 4;
 
     public static final int NUMBER_COLUMNS = 5;
-    private static final int NUMBEROFVISIBLEROWS = 8;
+    private static final int NUMBEROFVISIBLEROWS = 7;
 
     private final LayersTable grid;
     private final JPanel optionsPanelWrapper;
